+++ client/.eslintrc.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line no-undef
 module.exports = {
     parser: "@typescript-eslint/parser", // Specifies the ESLint parser
 
@@ -16,27 +17,42 @@ module.exports = {
             version: "detect", // Tells eslint-plugin-react to automatically detect the version of React to use
         },
     },
+    plugins: ["@typescript-eslint", "react"],
 
     extends: [
+        "eslint:recommended",
         "plugin:react/recommended", // Uses the recommended rules from @eslint-plugin-react
-
+        "plugin:@typescript-eslint/eslint-recommended",
         "plugin:@typescript-eslint/recommended", // Uses the recommended rules from @typescript-eslint/eslint-plugin
-
-        "prettier/@typescript-eslint", // Uses eslint-config-prettier to disable ESLint rules from @typescript-eslint/eslint-plugin that would conflict with prettier
-
-        "plugin:prettier/recommended", // Enables eslint-plugin-prettier and eslint-config-prettier. This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
     ],
 
     rules: {
         // Place to specify ESLint rules. Can be used to overwrite rules specified from the extended configs
         // e.g. "@typescript-eslint/explicit-function-return-type": "off",
-        "@typescript-eslint/no-empty-interface": "off",
-        "@typescript-eslint/no-empty-function": "off",
-        "@typescript-eslint/explicit-function-return-type": "off",
-        "@typescript-eslint/camelcase": "off",
-        "@typescript-eslint/class-name-casing": "off",
-        "@typescript-eslint/no-use-before-define": "off",
-        "prettier/prettier": "off",
-        "@typescript-eslint/explicit-module-boundary-types":"off"
+        "@typescript-eslint/no-empty-function": 1,
+        "@typescript-eslint/explicit-module-boundary-types": 0,
+        "react/jsx-no-bind": 1,
+        "react/destructuring-assignment": 1,
+        "react/no-unused-state": 1,
+        "no-case-declarations": 1,
+        "react/no-render-return-value": 1,
+        "react/no-unescaped-entities": 1,
+        "@typescript-eslint/no-empty-interface": 1,
+        "react/no-children-prop": 1,
+        "prefer-spread": 1,
+        "no-constant-condition": 1,
+        "@typescript-eslint/ban-types": [
+            1,
+            {
+                extendDefaults: true,
+                types: {
+                    "{}": false,
+                },
+            },
+        ],
+        "@typescript-eslint/ban-ts-comment": 1,
+        "prefer-rest-params": 1,
+        "react/display-name": 1,
+        "react/prop-types": 1,
     },
 };